oh sorry, not quite what I wanted. the existing new page functionality is good. I am talking about only for the share page functionality. here is my complete thought:
Great work! Now can you add an additional functionality to share pages? The share icon should either be in the navbar (just a share icon) or in the dashboard next to the delete button for each page. it should copy a link to clipboard and tell you to send it to whoever you want to give the page to. if anyone opens the link, it should show an almost normal page except without a hash. it should have some easy to see button to add this page, and if you click it, it should generate a hash and save it to the current browser's local storage. the link should probably include a json string of the localstorage or something like that. lemme know if you have any questions